+++ Gemfile
@@ -26,6 +26,9 @@ gem 'jbuilder', '~> 2.5'
 # gem 'redis', '~> 3.0'
 # Use ActiveModel has_secure_password
 gem 'bcrypt', '~> 3.1.7'
+gem 'knock'
+
+gem 'rack-cors', :require => 'rack/cors'
 
 gem 'rubocop', require: false
