+++ app/controllers/api/v1/base_controller.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1
+    class BaseController < ApplicationController
+      protect_from_forgery with: :null_session
+    end
+  end
+end
