+++ app/controllers/api/v1/users_controller.rb
@@ -1,0 +1,34 @@
+module Api
+  module V1
+    class UsersController < ApplicationController
+
+      def new
+        if current_user
+          flash[:notice] = 'You are already logged in'
+          redirect_to matches_path
+        else
+          @user = User.new
+        end
+      end
+
+      def create
+        @user = User.new(user_params)
+        if @user.save
+          session[:user_id] = @user.id
+          flash[:notice] = 'User created sucessfully'
+          redirect_to matches_path
+        else
+          flash[:alert] = 'Something wrong'
+          render :new
+        end
+      end
+
+      private
+
+      def user_params
+        params.require(:user).permit(:name, :email, :avatar, :username,
+                                     :password, :password_confirmation)
+      end
+    end
+  end
+end
