+++ app/controllers/application_controller.rb
@@ -1,7 +1,6 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
-
-  before_action :authorize
+  include Knock::Authenticable
 
   rescue_from ActiveRecord::RecordNotFound, :with => :render_404
   helper_method :current_user
@@ -14,12 +13,5 @@ class ApplicationController < ActionController::Base
 
   def current_user
     @current_user ||= User.find(session[:user_id]) if session[:user_id]
-  end
-
-  def authorize
-    unless current_user
-      flash[:notice] = "Log In first" if performed?
-      redirect_to log_in_path
-    end
   end
 end
