+++ app/controllers/user_token_controller.rb
@@ -1,0 +1,2 @@
+class UserTokenController < Knock::AuthTokenController
+end
