+++ config/application.rb
@@ -9,6 +9,13 @@ Bundler.require(*Rails.groups)
 module Matches
   class Application < Rails::Application
     config.active_job.queue_adapter = :sidekiq
+
+    config.middleware.use Rack::Cors do
+      allow do
+        origins '*'
+        resource '/*', :headers => :any, :methods => [:get, :post, :options]
+      end
+    end
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
